Added export SVG test spec
Had to change the way the new window was opened. It now opens to a blank
page and the SVG gets written to it as an img element after the window
is opened. I had to do this because I couldn't access the window's
elements when it had a data URI since it was a different protocol than
http and Chrome complained about protocol mismatch. Now they're both
http.